Split graph components into their own modules
