+++ src/views/TreasuryDashboard/components/Graph/ChartHelper.tsx
@@ -1,0 +1,9 @@
+import { Theme } from "@mui/material/styles";
+
+export const getTickStyle = (theme: Theme): Record<string, string | number> => {
+  return {
+    stroke: theme.palette.primary.light,
+    fill: theme.palette.primary.light,
+    strokeWidth: "0.1px",
+  };
+};
